# app/api/v1/payments/myfatoorah/session/route.ts
import {NextRequest,NextResponse} from 'next/server';
import {checkoutOrderSchema} from '@/lib/checkout-schema';
import {readAdminStore,writeAdminStore,type AdminOrderLine} from '@/lib/admin-store';
import {findEligiblePromotion,pricePromotion} from '@/lib/promotions';
import {bundleForQuantity,productLineTotal} from '@/lib/catalog';

export const runtime='nodejs';

export async function POST(request:NextRequest){
  const parsed=checkoutOrderSchema.safeParse(await request.json().catch(()=>null));
  if(!parsed.success||parsed.data.payment!=='knet')return NextResponse.json({code:'INVALID_ORDER',message:'Please complete every required checkout field.'},{status:400});
  const input=parsed.data;const token=process.env.MYFATOORAH_API_TOKEN;
  if(!token)return NextResponse.json({code:'PAYMENT_NOT_CONFIGURED',message:'MyFatoorah is awaiting merchant credentials.'},{status:503});
  const store=await readAdminStore();
  if(!store.settings.myFatoorahEnabled)return NextResponse.json({code:'PAYMENT_DISABLED',message:'MyFatoorah is currently unavailable.'},{status:409});
  let subtotal=0;const lines:AdminOrderLine[]=[];
  for(const line of input.lines){
    const product=store.products.find(item=>item.id===line.id&&item.status==='published');
    if(!product)return NextResponse.json({code:'PRODUCT_UNAVAILABLE',message:'A product in your bag is unavailable.'},{status:409});
    if(product.stock<line.qty)return NextResponse.json({code:'OUT_OF_STOCK',message:`Only ${product.stock} units of ${product.name} remain.`},{status:409});
    const lineTotal=productLineTotal(product,line.qty);const bundle=bundleForQuantity(product,line.qty);subtotal+=lineTotal;
    lines.push({productId:product.id,name:product.name,sku:product.id,size:line.size,color:product.color,quantity:line.qty,unitPrice:product.price,lineTotal,bundleLabel:bundle?.label,unitCost:product.cost,image:product.image});
  }
  const baseDelivery=subtotal>=store.settings.freeShippingFils?0:store.settings.standardDeliveryFils;
  const promotion=findEligiblePromotion(store,input.promotionCode,subtotal);const priced=pricePromotion(promotion,subtotal,baseDelivery);const total=subtotal-priced.discount+priced.delivery;
  const createdAt=new Date().toISOString();const id=crypto.randomUUID();const number=`GK-${createdAt.slice(2,10).replaceAll('-','')}-${String(store.orders.length+1).padStart(3,'0')}`;
  const methods=[store.settings.myFatoorahCardEnabled&&'card',store.settings.myFatoorahKnetEnabled&&'knet',store.settings.myFatoorahApplePayEnabled&&'applepay'].filter(Boolean) as string[];
  if(!methods.length)return NextResponse.json({code:'PAYMENT_METHODS_DISABLED',message:'No MyFatoorah payment methods are enabled.'},{status:409});
  const origin=request.nextUrl.origin;const apiUrl=process.env.MYFATOORAH_API_URL||'https://apitest.myfatoorah.com';
  const provider=await fetch(`${apiUrl}/v3/sessions`,{method:'POST',headers:{Authorization:`Bearer ${token}`,'Content-Type':'application/json','Idempotency-Key':input.idempotencyKey||id},body:JSON.stringify({PaymentMode:'COMPLETE_PAYMENT',Order:{Amount:total/1000,Currency:'KWD',ExternalIdentifier:id},Customer:{Name:`${input.firstName} ${input.lastName}`,Email:input.email,Mobile:input.mobile,Reference:id},SupportedPaymentMethods:methods,Language:input.locale==='ar'?'AR':'EN',IntegrationUrls:{Redirection:`${origin}/${input.locale}/checkout/myfatoorah/result`},MetaData:{orderId:id,orderNumber:number,promotionCode:priced.promotionCode||''}})});
  const result=await provider.json() as {IsSuccess?:boolean;Message?:string;Data?:{SessionId?:string}};const sessionId=result.Data?.SessionId;
  if(!provider.ok||!result.IsSuccess||!sessionId)return NextResponse.json({code:'MYFATOORAH_ERROR',message:result.Message||'MyFatoorah could not create a payment session.'},{status:502});
  const order={id,number,customer:`${input.firstName} ${input.lastName}`,email:input.email,mobile:input.mobile,total,subtotal,delivery:priced.delivery,discount:priced.discount,promotionId:priced.promotionId,promotionCode:priced.promotionCode,payment:'knet' as const,paymentStatus:'pending' as const,status:'pending' as const,createdAt,address:{area:input.area,block:input.block,street:input.street,building:input.building,floor:input.floor},lines,myFatoorahSessionId:sessionId,idempotencyKey:input.idempotencyKey,inventoryReleased:false};
  const products=store.products.map(product=>{const line=lines.find(item=>item.productId===product.id);return line?{...product,stock:product.stock-line.quantity}:product});
  const promotions=promotion?store.promotions.map(item=>item.id===promotion.id?{...item,uses:(item.uses??0)+1}:item):store.promotions;
  const existing=store.customers.find(customer=>customer.email.toLowerCase()===input.email.toLowerCase());
  const customers=existing?store.customers.map(customer=>customer.id===existing.id?{...customer,name:order.customer,mobile:input.mobile,orders:customer.orders+1,lastOrderAt:createdAt}:customer):[{id:crypto.randomUUID(),name:order.customer,email:input.email,mobile:input.mobile,orders:1,spent:0,createdAt,lastOrderAt:createdAt,tags:['MyFatoorah customer']},...store.customers];
  await writeAdminStore({...store,products,promotions,customers,orders:[order,...store.orders]},'Created MyFatoorah payment session',number);
  return NextResponse.json({data:{number,sessionId}});
}