# app/api/v1/payments/stripe/session/route.ts
import {NextRequest,NextResponse} from 'next/server';
import {checkoutOrderSchema} from '@/lib/checkout-schema';
import {nextOrderNumber,readAdminStore,writeAdminStore,type AdminOrderLine} from '@/lib/admin-store';
import {findEligiblePromotion,pricePromotion} from '@/lib/promotions';
import {bundleForQuantity,productLineTotal} from '@/lib/catalog';
import {getStripe} from '@/lib/stripe';

export const runtime='nodejs';

export async function POST(request:NextRequest){
  const parsed=checkoutOrderSchema.safeParse(await request.json().catch(()=>null));
  if(!parsed.success||parsed.data.payment!=='stripe')return NextResponse.json({code:'INVALID_ORDER',message:'Please complete every required checkout field.'},{status:400});
  const input=parsed.data;const store=await readAdminStore();
  if(!store.settings.stripeEnabled)return NextResponse.json({code:'STRIPE_DISABLED',message:'Stripe is not currently enabled.'},{status:409});
  if(!process.env.STRIPE_SECRET_KEY)return NextResponse.json({code:'STRIPE_NOT_CONFIGURED',message:'Stripe is awaiting merchant connection.'},{status:503});
  if(input.idempotencyKey){const existing=store.orders.find(order=>order.idempotencyKey===input.idempotencyKey&&order.stripeSessionId);if(existing?.stripeCheckoutUrl)return NextResponse.json({data:{number:existing.number,url:existing.stripeCheckoutUrl}})}
  let subtotal=0;const lines:AdminOrderLine[]=[];
  for(const line of input.lines){
    const product=store.products.find(item=>item.id===line.id&&item.status==='published');
    if(!product)return NextResponse.json({code:'PRODUCT_UNAVAILABLE',message:'A product in your bag is no longer available.'},{status:409});
    if(product.stock<line.qty)return NextResponse.json({code:'OUT_OF_STOCK',message:`Only ${product.stock} units of ${product.name} remain.`},{status:409});
    const lineTotal=productLineTotal(product,line.qty);const bundle=bundleForQuantity(product,line.qty);subtotal+=lineTotal;
    lines.push({productId:product.id,name:product.name,sku:product.id,size:line.size,color:product.color,quantity:line.qty,unitPrice:product.price,lineTotal,bundleLabel:bundle?.label,unitCost:product.cost,image:product.image});
  }
  const baseDelivery=subtotal>=store.settings.freeShippingFils?0:store.settings.standardDeliveryFils;
  const promotion=findEligiblePromotion(store,input.promotionCode,subtotal);const priced=pricePromotion(promotion,subtotal,baseDelivery);const total=subtotal-priced.discount+priced.delivery;
  const createdAt=new Date().toISOString();const id=crypto.randomUUID();const number=nextOrderNumber(store.orders,createdAt);const origin=request.nextUrl.origin;
  try{
    const stripe=getStripe();const factor=promotion?.discount?1-promotion.discount/100:1;
    const session=await stripe.checkout.sessions.create({mode:'payment',customer_email:input.email,client_reference_id:id,success_url:`${origin}/${input.locale}/checkout/success?session_id={CHECKOUT_SESSION_ID}`,cancel_url:`${origin}/${input.locale}/checkout?payment=cancelled`,expires_at:Math.floor(Date.now()/1000)+31*60,billing_address_collection:store.settings.stripeBillingAddress?'required':'auto',allow_promotion_codes:store.settings.stripeAllowPromotionCodes,line_items:[...lines.map(line=>({quantity:1,price_data:{currency:'kwd',unit_amount:Math.round((line.lineTotal??line.unitPrice*line.quantity)*factor),product_data:{name:`${line.name} × ${line.quantity}`,description:[line.color,line.size,line.bundleLabel].filter(Boolean).join(' · '),metadata:{productId:line.productId,sku:line.sku}}}})),...(priced.delivery?[{quantity:1,price_data:{currency:'kwd',unit_amount:priced.delivery,product_data:{name:'Kuwait standard delivery'}}}]:[])],metadata:{orderId:id,orderNumber:number,promotionCode:priced.promotionCode||''},payment_intent_data:{metadata:{orderId:id,orderNumber:number,promotionCode:priced.promotionCode||''},receipt_email:input.email},customer_creation:'always'},{idempotencyKey:input.idempotencyKey});
    if(!session.url)throw new Error('Stripe did not return a checkout URL');
    const order={id,number,customer:`${input.firstName} ${input.lastName}`,email:input.email,mobile:input.mobile,total,subtotal,delivery:priced.delivery,discount:priced.discount,promotionId:priced.promotionId,promotionCode:priced.promotionCode,payment:'stripe' as const,paymentStatus:'pending' as const,status:'pending' as const,createdAt,address:{area:input.area,block:input.block,street:input.street,building:input.building,floor:input.floor},lines,stripeSessionId:session.id,stripeCheckoutUrl:session.url,idempotencyKey:input.idempotencyKey,inventoryReleased:false};
    const products=store.products.map(product=>{const line=lines.find(item=>item.productId===product.id);return line?{...product,stock:product.stock-line.quantity}:product});
    const promotions=promotion?store.promotions.map(item=>item.id===promotion.id?{...item,uses:(item.uses??0)+1}:item):store.promotions;
    const existing=store.customers.find(customer=>customer.email.toLowerCase()===input.email.toLowerCase());
    const customers=existing?store.customers.map(customer=>customer.id===existing.id?{...customer,name:order.customer,mobile:input.mobile,orders:customer.orders+1,lastOrderAt:createdAt}:customer):[{id:crypto.randomUUID(),name:order.customer,email:input.email,mobile:input.mobile,orders:1,spent:0,createdAt,lastOrderAt:createdAt,tags:['Stripe customer']},...store.customers];
    await writeAdminStore({...store,products,promotions,customers,orders:[order,...store.orders]},'Created Stripe Checkout session',number);
    return NextResponse.json({data:{number,url:session.url,sessionId:session.id}});
  }catch{return NextResponse.json({code:'STRIPE_ERROR',message:'Stripe checkout is temporarily unavailable. Please choose another payment method.'},{status:502})}
}