# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "GIANT Commerce API", "version": "1.2.0"},
  "servers": [{"url": "/api/v1"}],
  "paths": {
    "/catalog/products": {"get": {"summary": "List and search products", "responses": {"200": {"description": "Product cursor page"}}}},
    "/catalog/categories": {"get": {"summary": "List categories", "responses": {"200": {"description": "Categories"}}}},
    "/catalog/products/{slug}/size-guide": {"get": {"summary": "Get the published size guide assigned to a product", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Product size guide"}, "404": {"description": "Product not found"}}}},
    "/size-assistant/turn": {"post": {"summary": "Continue a stateless bilingual sizing conversation", "responses": {"200": {"description": "Assistant turn and optional recommendation"}, "429": {"description": "Rate limited"}}}},
    "/checkout/quote": {"post": {"summary": "Calculate an authoritative checkout quote", "responses": {"200": {"description": "Quote"}}}},
    "/payments/myfatoorah/session": {"post": {"summary": "Create a MyFatoorah v3 payment session", "responses": {"200": {"description": "Payment session"}, "503": {"description": "Provider not configured"}}}},
    "/payments/stripe/session": {"post": {"summary": "Create a Stripe Checkout session", "responses": {"200": {"description": "Hosted checkout URL"}, "503": {"description": "Provider not configured"}}}},
    "/payments/tap/session": {"post": {"summary": "Create a Tap hosted charge", "responses": {"200": {"description": "Hosted checkout URL"}, "503": {"description": "Provider not configured"}}}},
    "/meta/events": {"post": {"summary": "Record a browser commerce event and forward it through Meta CAPI", "responses": {"202": {"description": "Accepted"}, "400": {"description": "Invalid event"}}}},
    "/webhooks/myfatoorah": {"post": {"summary": "Receive signed MyFatoorah payment events", "responses": {"200": {"description": "Accepted"}, "401": {"description": "Invalid signature"}}}},
    "/webhooks/stripe": {"post": {"summary": "Receive signed Stripe events", "responses": {"200": {"description": "Accepted"}, "401": {"description": "Invalid signature"}}}},
    "/webhooks/tap": {"post": {"summary": "Receive Tap events and re-fetch the authoritative charge", "responses": {"200": {"description": "Accepted"}, "502": {"description": "Provider verification failed"}}}}
  },
  "components": {"schemas": {
    "Money": {"type": "object", "required": ["amount", "currency"], "properties": {"amount": {"type": "integer", "description": "Integer fils"}, "currency": {"const": "KWD"}}},
    "ApiError": {"type": "object", "required": ["code", "message", "requestId"], "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "fieldErrors": {"type": "object"}, "requestId": {"type": "string"}}},
    "SizingRecommendation": {"type": "object", "required": ["productId", "sizeGuideVersion", "recommendedSize", "confidence", "reasons", "missingMeasurements", "status"], "properties": {"productId": {"type": "string"}, "sizeGuideVersion": {"type": "string"}, "recommendedSize": {"type": ["string", "null"]}, "alternateSize": {"type": "string"}, "confidence": {"enum": ["high", "medium", "low"]}, "reasons": {"type": "array", "items": {"type": "string"}}, "missingMeasurements": {"type": "array", "items": {"type": "string"}}, "status": {"enum": ["collecting", "recommended", "cannot_recommend"]}}}
  }}
}